Module jss::common

Modules

unit

Structs

Context
Context with other needed info - for parse and prepares, aka dimensions screen, element measures, variables, and other.

ParseOptions
Options for parse style from uniform functions

Style
Style element, with all element status, and context`s, with implementations of traits for parse unions of one element

StyleProperties
Style properties struct with all css properties, by default allow to raw serialize/deserialize with serde. All elements wrap over Option for support optional field. Also implements derive macro for Add operator for merge properties, and custom macros for relative parse & prepares for styles

Enums

ParseError

ParseTarget
From what data format to parse, currently available: JSON, YAML Default: JSON

PropertyCase
What format of properties keys to use and check when parsing Current allowed: snake_case, camelCase, kebab-case Default: camelCase

RecursiveType
One level recursion inside with: @media queries, status presudo classes: :hover, :active. Default: Basic

Traits

ParseStyleExt

PrepareStyleExt