Struct jss::common::StyleProperties

pub struct StyleProperties {
    pub flex_direction: Option<FlexDirection>,
    pub justify_content: Option<Justify>,
    pub position: Option<PositionType>,
    pub align_content: Option<Align>,
    pub align_items: Option<Align>,
    pub align_self: Option<Align>,
    pub flex_wrap: Option<Wrap>,
    pub display: Option<Display>,
    pub overflow: Option<Overflow>,
    pub aspect_ratio: Option<OrderedFloat<f32>>,
    pub flex_shrink: Option<OrderedFloat<f32>>,
    pub flex_grow: Option<OrderedFloat<f32>>,
    pub flex: Option<OrderedFloat<f32>>,
    pub bottom: Option<StyleUnit>,
    pub end: Option<StyleUnit>,
    pub flex_basis: Option<StyleUnit>,
    pub height: Option<StyleUnit>,
    pub left: Option<StyleUnit>,
    pub margin: Option<StyleUnit>,
    pub margin_bottom: Option<StyleUnit>,
    pub margin_end: Option<StyleUnit>,
    pub margin_horizontal: Option<StyleUnit>,
    pub margin_left: Option<StyleUnit>,
    pub margin_right: Option<StyleUnit>,
    pub margin_start: Option<StyleUnit>,
    pub margin_top: Option<StyleUnit>,
    pub margin_vertical: Option<StyleUnit>,
    pub max_height: Option<StyleUnit>,
    pub max_width: Option<StyleUnit>,
    pub min_height: Option<StyleUnit>,
    pub min_width: Option<StyleUnit>,
    pub padding: Option<StyleUnit>,
    pub padding_bottom: Option<StyleUnit>,
    pub padding_end: Option<StyleUnit>,
    pub padding_horizontal: Option<StyleUnit>,
    pub padding_left: Option<StyleUnit>,
    pub padding_right: Option<StyleUnit>,
    pub padding_start: Option<StyleUnit>,
    pub padding_top: Option<StyleUnit>,
    pub padding_vertical: Option<StyleUnit>,
    pub right: Option<StyleUnit>,
    pub start: Option<StyleUnit>,
    pub top: Option<StyleUnit>,
    pub width: Option<StyleUnit>,
    pub background: Option<Background>,
    pub transform: Option<Transforms>,
    pub filter: Option<Filters>,
    pub border_bottom_width: Option<OrderedFloat<f32>>,
    pub border_right_width: Option<OrderedFloat<f32>>,
    pub border_left_width: Option<OrderedFloat<f32>>,
    pub border_top_width: Option<OrderedFloat<f32>>,
    pub border_top_color: Option<Color>,
    pub border_right_color: Option<Color>,
    pub border_left_color: Option<Color>,
    pub border_bottom_color: Option<Color>,
    pub border_top_style: Option<BorderStyle>,
    pub border_right_style: Option<BorderStyle>,
    pub border_left_style: Option<BorderStyle>,
    pub border_bottom_style: Option<BorderStyle>,
    pub border_top_right_radius: Option<i32>,
    pub border_top_left_radius: Option<i32>,
    pub border_bottom_right_radius: Option<i32>,
    pub border_bottom_left_radius: Option<i32>,
}

Style properties struct with all css properties, by default allow to raw serialize/deserialize with serde. All elements wrap over Option for support optional field. Also implements derive macro for Add operator for merge properties, and custom macros for relative parse & prepares for styles

Fields
flex_direction: Option<FlexDirection> justify_content: Option<Justify> position: Option<PositionType> align_content: Option<Align> align_items: Option<Align> align_self: Option<Align> flex_wrap: Option<Wrap> display: Option<Display> overflow: Option<Overflow> aspect_ratio: Option<OrderedFloat<f32>> flex_shrink: Option<OrderedFloat<f32>> flex_grow: Option<OrderedFloat<f32>> flex: Option<OrderedFloat<f32>> bottom: Option<StyleUnit> end: Option<StyleUnit> flex_basis: Option<StyleUnit> height: Option<StyleUnit> left: Option<StyleUnit> margin: Option<StyleUnit> margin_bottom: Option<StyleUnit> margin_end: Option<StyleUnit> margin_horizontal: Option<StyleUnit> margin_left: Option<StyleUnit> margin_right: Option<StyleUnit> margin_start: Option<StyleUnit> margin_top: Option<StyleUnit> margin_vertical: Option<StyleUnit> max_height: Option<StyleUnit> max_width: Option<StyleUnit> min_height: Option<StyleUnit> min_width: Option<StyleUnit> padding: Option<StyleUnit> padding_bottom: Option<StyleUnit> padding_end: Option<StyleUnit> padding_horizontal: Option<StyleUnit> padding_left: Option<StyleUnit> padding_right: Option<StyleUnit> padding_start: Option<StyleUnit> padding_top: Option<StyleUnit> padding_vertical: Option<StyleUnit> right: Option<StyleUnit> start: Option<StyleUnit> top: Option<StyleUnit> width: Option<StyleUnit> background: Option<Background> transform: Option<Transforms> filter: Option<Filters> border_bottom_width: Option<OrderedFloat<f32>> border_right_width: Option<OrderedFloat<f32>> border_left_width: Option<OrderedFloat<f32>> border_top_width: Option<OrderedFloat<f32>> border_top_color: Option<Color> border_right_color: Option<Color> border_left_color: Option<Color> border_bottom_color: Option<Color> border_top_style: Option<BorderStyle> border_right_style: Option<BorderStyle> border_left_style: Option<BorderStyle> border_bottom_style: Option<BorderStyle> border_top_right_radius: Option<i32> border_top_left_radius: Option<i32> border_bottom_right_radius: Option<i32> border_bottom_left_radius: Option<i32>