Trait jss::traits::TStyleContext

pub trait TStyleContext: Debug + PartialEq + Clone {
    fn set_dimensions(&mut self, _: Option<Dimensions>);

    fn set_variable(&mut self, _: String, _: Variable);

    fn set_variables<T>(&mut self, _: T)
    where
        T: IntoIterator<Item = (String, Variable)>;
}

Manipulate runtime context of current style - work with active element dimensions, set variables for runtime calculation and other data needed in runtime

Required Methods

fn set_dimensions(&mut self, _: Option<Dimensions>)

Set dimensions with size and position node on layout

fn set_variable(&mut self, _: String, _: Variable)

Set variables for runtime calculator after apply you can use that var inside calc expression

fn set_variables<T>(&mut self, _: T) where
T: IntoIterator<Item = (String, Variable)>,

The same thing as set_variable but for enumeration items