Struct jss::types::style_properties_repr::StyleProperties

pub struct StyleProperties {
    pub flex_direction: Option<FlexDirection>,
    pub justify_content: Option<Justify>,
    pub position: Option<PositionType>,
    pub align_content: Option<Align>,
    pub align_items: Option<Align>,
    pub align_self: Option<Align>,
    pub flex_wrap: Option<Wrap>,
    pub display: Option<Display>,
    pub overflow: Option<Overflow>,
    pub aspect_ratio: Option<AspectRatio>,
    pub flex_shrink: Option<FlexShrink>,
    pub flex_grow: Option<FlexGrow>,
    pub flex: Option<FlexFactor>,
    pub bottom: Option<SharedUnit>,
    pub end: Option<SharedUnit>,
    pub flex_basis: Option<SharedUnit>,
    pub height: Option<SharedUnit>,
    pub left: Option<SharedUnit>,
    pub margin: Option<SharedUnit>,
    pub margin_bottom: Option<SharedUnit>,
    pub margin_end: Option<SharedUnit>,
    pub margin_horizontal: Option<SharedUnit>,
    pub margin_left: Option<SharedUnit>,
    pub margin_right: Option<SharedUnit>,
    pub margin_start: Option<SharedUnit>,
    pub margin_top: Option<SharedUnit>,
    pub margin_vertical: Option<SharedUnit>,
    pub max_height: Option<SharedUnit>,
    pub max_width: Option<SharedUnit>,
    pub min_height: Option<SharedUnit>,
    pub min_width: Option<SharedUnit>,
    pub padding: Option<SharedUnit>,
    pub padding_bottom: Option<SharedUnit>,
    pub padding_end: Option<SharedUnit>,
    pub padding_horizontal: Option<SharedUnit>,
    pub padding_left: Option<SharedUnit>,
    pub padding_right: Option<SharedUnit>,
    pub padding_start: Option<SharedUnit>,
    pub padding_top: Option<SharedUnit>,
    pub padding_vertical: Option<SharedUnit>,
    pub right: Option<SharedUnit>,
    pub start: Option<SharedUnit>,
    pub top: Option<SharedUnit>,
    pub width: Option<SharedUnit>,
    pub border_bottom_width: Option<BorderWidth>,
    pub border_right_width: Option<BorderWidth>,
    pub border_left_width: Option<BorderWidth>,
    pub border_top_width: Option<BorderWidth>,
    pub background: Option<Background>,
    pub transform: Option<Transforms>,
    pub filter: Option<Filters>,
    pub border_top_color: Option<BorderColor>,
    pub border_right_color: Option<BorderColor>,
    pub border_left_color: Option<BorderColor>,
    pub border_bottom_color: Option<BorderColor>,
    pub border_top_style: Option<BorderStyle>,
    pub border_right_style: Option<BorderStyle>,
    pub border_left_style: Option<BorderStyle>,
    pub border_bottom_style: Option<BorderStyle>,
    pub border_top_right_radius: Option<BorderRadius>,
    pub border_top_left_radius: Option<BorderRadius>,
    pub border_bottom_right_radius: Option<BorderRadius>,
    pub border_bottom_left_radius: Option<BorderRadius>,
}

Style properties struct with all css properties, by default allow to raw serialize/deserialize with serde. All elements wrap over Option for support optional field. Also implements derive macro for Add operator for merge properties, and custom macros for relative parse & prepares for styles

Fields
flex_direction: Option<FlexDirection> justify_content: Option<Justify> position: Option<PositionType> align_content: Option<Align> align_items: Option<Align> align_self: Option<Align> flex_wrap: Option<Wrap> display: Option<Display> overflow: Option<Overflow> aspect_ratio: Option<AspectRatio> flex_shrink: Option<FlexShrink> flex_grow: Option<FlexGrow> flex: Option<FlexFactor> bottom: Option<SharedUnit> end: Option<SharedUnit> flex_basis: Option<SharedUnit> height: Option<SharedUnit> left: Option<SharedUnit> margin: Option<SharedUnit> margin_bottom: Option<SharedUnit> margin_end: Option<SharedUnit> margin_horizontal: Option<SharedUnit> margin_left: Option<SharedUnit> margin_right: Option<SharedUnit> margin_start: Option<SharedUnit> margin_top: Option<SharedUnit> margin_vertical: Option<SharedUnit> max_height: Option<SharedUnit> max_width: Option<SharedUnit> min_height: Option<SharedUnit> min_width: Option<SharedUnit> padding: Option<SharedUnit> padding_bottom: Option<SharedUnit> padding_end: Option<SharedUnit> padding_horizontal: Option<SharedUnit> padding_left: Option<SharedUnit> padding_right: Option<SharedUnit> padding_start: Option<SharedUnit> padding_top: Option<SharedUnit> padding_vertical: Option<SharedUnit> right: Option<SharedUnit> start: Option<SharedUnit> top: Option<SharedUnit> width: Option<SharedUnit> border_bottom_width: Option<BorderWidth> border_right_width: Option<BorderWidth> border_left_width: Option<BorderWidth> border_top_width: Option<BorderWidth> background: Option<Background> transform: Option<Transforms> filter: Option<Filters> border_top_color: Option<BorderColor> border_right_color: Option<BorderColor> border_left_color: Option<BorderColor> border_bottom_color: Option<BorderColor> border_top_style: Option<BorderStyle> border_right_style: Option<BorderStyle> border_left_style: Option<BorderStyle> border_bottom_style: Option<BorderStyle> border_top_right_radius: Option<BorderRadius> border_top_left_radius: Option<BorderRadius> border_bottom_right_radius: Option<BorderRadius> border_bottom_left_radius: Option<BorderRadius>